# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FonAd / DAADD API
  version: 1.0.0
  description: |
    FonAd (DAADD) is a full-featured digital advertising and analytics platform.
    It provides campaign management, real-time event tracking, AI-driven optimization,
    anomaly detection, cross-device attribution, benchmarking, storytelling reports,
    viewer reward programs, team collaboration, and notification services.
  contact:
    name: FonAd Engineering
    email: engineering@fonad.io
servers:
  - url: http://localhost:4000/api/v1
    description: Local development server
security:
  - BearerAuth: []
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      description: Creates a new user account and returns a JWT token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in
      description: Authenticates a user by email and password and returns a JWT token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Account is deactivated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user
      description: Returns the profile of the currently authenticated user.
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ─────────────────────────── Users ───────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login or /auth/register
...
  responses:
    BadRequest:
      description: Bad request -- missing or invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized -- missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    UserRole:
      type: string
      enum: [advertiser, campaign_manager, analyst, end_user, admin]
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, email, name, role, is_active, created_at, updated_at]
...
    RegisterRequest:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
          description: Must be at least 8 characters
        name:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
          description: JWT bearer token
      required: [user, token]
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          description: Response payload (type varies by endpoint)
      required: [success]
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
              description: Human-readable error message
            code:
              type: string
              description: Machine-readable error code
              enum:
                - BAD_REQUEST
                - UNAUTHORIZED
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - INTERNAL_ERROR
                - INVALID_CREDENTIALS
                - EMAIL_EXISTS
                - ACCOUNT_INACTIVE
                - INVALID_TRANSITION
                - INVALID_BUDGET
                - AI_NOT_READY
            details:
              description: Additional error context
          required: [message, code]
      required: [success, error]